Extended this PR by: - adding restat of the script after installation of dependencies - adding session to connectionGlobals - renaming GetListOfOpendModels to GetListOfOpenedModels - putting first basic example into new folder (BasicExample)
